==> 1_HBR headlines exploration and cleanup.ipynb <==
import pandas as pd
import numpy as np

from nltk import word_tokenize
from nltk import FreqDist
from nltk.corpus import stopwords
from nltk.tokenize import RegexpTokenizer
import nltk

## Let's explore our headlines

#Load headlines and see a few
data = pd.read_csv("headlines.csv")
print data.shape
data.head()

#Some text functions

#Takes a list of words, remove common ones
def remove_stop_words(tokens):
    stopwords = nltk.corpus.stopwords.words('english')
    content = [w for w in tokens if w.lower() not in stopwords]
    return content

#Takes a list of words, lowercases them
def lowercase_tokens(tokens):
    list_of_words = []
    for a in tokens:
        a = a.lower()
        list_of_words.append(a)
    return list_of_words

#Count the most common words in headlines

#Concatenate headlines into one string
text = data['Page Title'].str.cat(sep=" ")

#Encode as utf-8 
text = text.decode("utf-8")

#Create tokenizer to remove punctuation and numbers
tokenizer = RegexpTokenizer(r'\w+')

#Tokenize into words, lowercase, remove stop words
tokens = tokenizer.tokenize(text)
tokens = lowercase_tokens(tokens)
tokens = remove_stop_words(tokens)

#Print common words
fdist = FreqDist(tokens)
print "The most common words in our headlines are:"
print fdist.most_common(50)


## The exploration showed that lots of headlines have "HBR" in them. We need to fix that, look for similar issues, and then get rid of duplicates

#Subset headlines that have "HBR" and other issues in them to see how they look

hbr = data[data['Page Title'].str.contains("HBR")]
print hbr.shape

hbr2 = data[data['Page Title'].str.contains("Harvard Business Review")]
print hbr2.shape

for a in hbr['Page Title'][0:20]:
    print a
    

#Fix headline formatting and remove duplicates
data['Clean Title'] = data['Page Title'].str.replace(' - HBR', '')

#Check for remaining 'HBR' mentions
hbr = data[data['Clean Title'].str.contains("HBR")]
print hbr.shape
hbr.head()

#We've cut 400+ headlines down to 15. Now let's just remove "HBR" from those remaining heds
data['Clean Title'] = data['Clean Title'].str.replace('HBR', '')

#Drop duplicate headlines
data = data.drop_duplicates(subset="Clean Title")
data.shape


## We've got ~4500 cleaned up headlines left. Time to save them to a new csv

#data['Clean Title'].to_csv("clean_headlines.csv",header=True,index=False)

## To add to the dataset, add social media posts from Twitter & Facebook

social = pd.read_csv("all_tweets_fb_2013-nov18.csv")
print social.shape
print social.columns
social['Created By'].describe()

#Limit the dataset to tweets and Facebook posts by HBR editors, as opposed to marketing, etc.

editors = ['nicole.torres@harvardbusiness.org','alexandra.kephart@hbr.org', 'Ramsey.Khabbaz@harvardbusiness.org',
     'paige.cohen@hbr.org','nicole.blank@hbr.org','ggavett@hbr.org','etruxler@hbr.org','awieckowski@hbr.org',
     'duygu.mullin@hbr.org','walter.frick@harvardbusiness.org']
social = social[social['Created By'].isin(editors)]
social.shape

social['Message'].head

#Combine text fields for HBR headlines and social media

tk = data['Clean Title']
tk2 = social['Message']

frames = [tk,tk2]

result = pd.concat(frames)
df = result.to_frame(name='Clean Title')
print df.shape
df.head()

#Remove mentions of HBR
df['Clean Title'] = df['Clean Title'].str.replace(' - HBR', '')
df['Clean Title'] = df['Clean Title'].str.replace('HBR', '')

#Remove mentions of research because the algorithm is being used to evaluate research
df['Clean Title'] = df['Clean Title'].str.replace('Research', '')
df['Clean Title'] = df['Clean Title'].str.replace('research', '')
df.head()


## Save the combined dataframe to csv

df['Clean Title'].to_csv("clean_headlines.csv",header=True,index=False)
