==> 2_Headline aggregator exploration.ipynb <==
import pandas as pd
import numpy as np

#Load headline dataset and take a look
# https://www.kaggle.com/uciml/news-aggregator-dataset

df = pd.read_csv("uci-news-aggregator.csv")
print df.shape

df.head()


#Randomly select 15,000 headlines (to match size of HBR dataset)
random_df = df.sample(n=15000)

#Save headlines from random set to csv
random_df['TITLE'].to_csv("other_headlines.csv",index=False,header=True)

#Do the same just for business headlines
biz = df.loc[df['CATEGORY'] == 'b']
random_biz = biz.sample(n=15000)

#Save random business headlines to cvs
random_biz['TITLE'].to_csv("other_biz_headlines.csv",index=False,header=True)
