==> 3_combine data.ipynb <==
import numpy as np
import pandas as pd

## Combine clean HBR headlines with sampling of aggregator headlines and add labels

hbr = pd.read_csv("clean_headlines.csv")
hbr.columns = ['Headline']
hbr['HBR'] = 'Yes'
print hbr.shape
hbr.head()

others = pd.read_csv("other_headlines.csv")
others.columns = ['Headline']
others['HBR'] = 'No'
print others.shape
others.head()

combined = pd.concat([hbr, others])
print combined.shape
combined.head()

combined.to_csv("combined_headlines.csv",index=False,header=True)

## Repeat for business headlines only

biz = pd.read_csv("other_biz_headlines.csv")
biz.columns = ['Headline']
biz['HBR'] = 'No'
print biz.shape
biz.head()

combined_biz = pd.concat([hbr, biz])
print combined_biz.shape
combined_biz.head()

combined_biz.to_csv("combined_biz_headlines.csv",index=False,header=True)
