==> 4_train bag of words model.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn import linear_model
from sklearn.metrics import accuracy_score
import pickle
from sklearn.pipeline import Pipeline

## Load data and split into train/test

#Load and check data
heds = pd.read_csv("combined_headlines.csv")
print heds.shape
print heds.describe()
heds.head()

#Split data into train/test
train = heds.sample(frac=0.7, random_state=1)
test = heds.loc[~heds.index.isin(train.index)]

print "Train shape:"
print train.shape
print "Test shape"
print test.shape

## Vectorize and train model

#Vectorize data

#Initialize countvectorizer and fit to headlines
vectorizer = CountVectorizer(analyzer='word',
                             stop_words = 'english',
                             ngram_range=(1,2),
                             max_features=1000)

train_counts = vectorizer.fit_transform(train['Headline'])
test_counts = vectorizer.transform(test['Headline'])

#Initialize tfidf transformer and fit to counts
tfidf_transformer = TfidfTransformer()
train_tfidf = tfidf_transformer.fit_transform(train_counts)
print train_tfidf.shape

test_tfidf = tfidf_transformer.transform(test_counts)
print test_tfidf.shape

#Train and evaluate linear model

#Train logistic regression model on training set
logit = linear_model.LogisticRegression(penalty='l1',C=1)
logit = logit.fit(train_tfidf,train['HBR'])

#Evaluate training performance
logit_train_results = logit.predict(train_tfidf)
train_output = pd.DataFrame(data={"Headline":train['Headline'],
                                 "HBR":train["HBR"],
                                 "Prediction":logit_train_results})
print train_output.head(30)

print "\nThe accuracy of the model on the training set is:"
print accuracy_score(train_output['HBR'],train_output['Prediction'])

#Make predictions against test data and assess performance
logit_test_results = logit.predict(test_tfidf)
test_output = pd.DataFrame(data={"Headline":test['Headline'],
                                 "HBR":test["HBR"],
                                 "Prediction":logit_test_results})
print test_output.head(30)

print "\nThe accuracy of the model on the test set is:"
print accuracy_score(test_output['HBR'],test_output['Prediction'])

## Save the trained model

with open('hbr_logit.pkl','wb') as f:
    pickle.dump(logit,f)

#Turn text into dataframe to test
text = "How to manage a company"   
text = [text]
df = pd.DataFrame(text,columns =['text'])    
    
#Vectorize text
tk_features = vectorizer.transform(df['text'])
tk_features = tfidf_transformer.transform(tk_features)
    
#Return prediction
result = logit.predict(tk_features)
print result[0]

## Redo the same process, but using a pipeline

#Pipeline makes it easier to vectorize new data later on -- to save the vectorizer along with the model

vect = CountVectorizer(analyzer='word',
                             stop_words = 'english',
                             ngram_range=(1,2),
                             max_features=1000)

#Define pipeline with countvectorizer, tfidftransformer, and logit model
test_pipe = Pipeline([
     ('vectorizer', vect),
     ('tfidf', TfidfTransformer()),
     ('logit', linear_model.LogisticRegression(penalty='l1',C=1))
 ])

#Fit logistic regression model with training data
test_pipe.fit(train['Headline'], train["HBR"]) 

#Make predictions against test data and assess performance
predictions = test_pipe.predict(test['Headline'])
test_output = pd.DataFrame(data={"Headline":test['Headline'],
                                 "HBR":test["HBR"],
                                 "Prediction":predictions})
print test_output.head(30)

print "\nThe accuracy of the model on the test set is:"
print accuracy_score(test_output['HBR'],test_output['Prediction'])


#Print the features with the lowest and highest coefficients
feature_names = np.array(vect.get_feature_names())

sorted_coef_index = logit.coef_[0].argsort()
print('Smallest Coefs: \n{}\n'.format(feature_names[sorted_coef_index[:100]]))
print('Largest Coefs: \n{}\n'.format(feature_names[sorted_coef_index[:-101:-1]]))


#Test new text (defined above) with fitted model

print type(text)
print type(test['Headline'])

print test_pipe.predict(pd.Series(text))

#Save pipeline for later

with open('hbr_pipeline.pkl','wb') as f:
    pickle.dump(test_pipe,f)

# Repeat pipeline building and training with business heds only

#Load and check data
heds = pd.read_csv("combined_biz_headlines.csv")
print heds.shape
print heds.describe()
heds.head()

#Split data into train/test
train = heds.sample(frac=0.7, random_state=1)
test = heds.loc[~heds.index.isin(train.index)]

print "Train shape:"
print train.shape
print "Test shape"
print test.shape

#Pipeline makes it easier to vectorize new data later on -- to save the vectorizer along with the model

vect = CountVectorizer(analyzer='word',
                             stop_words = 'english',
                             ngram_range=(1,2),
                             max_features=1000)

#Define pipeline with countvectorizer, tfidftransformer, and logit model
test_pipe = Pipeline([
     ('vectorizer', vect),
     ('tfidf', TfidfTransformer()),
     ('logit', linear_model.LogisticRegression(penalty='l1',C=1))
 ])

#Fit logistic regression model with training data
test_pipe.fit(train['Headline'], train["HBR"]) 

#Make predictions against test data and assess performance
predictions = test_pipe.predict(test['Headline'])
test_output = pd.DataFrame(data={"Headline":test['Headline'],
                                 "HBR":test["HBR"],
                                 "Prediction":predictions})
print test_output.head(30)

print "\nThe accuracy of the model on the test set is:"
print accuracy_score(test_output['HBR'],test_output['Prediction'])


#Print the features with the lowest and highest coefficients
feature_names = np.array(vect.get_feature_names())

sorted_coef_index = logit.coef_[0].argsort()
print('Smallest Coefs: \n{}\n'.format(feature_names[sorted_coef_index[:100]]))
print('Largest Coefs: \n{}\n'.format(feature_names[sorted_coef_index[:-101:-1]]))


#Test new text (defined above) with fitted model

print type(text)
print type(test['Headline'])

print test_pipe.predict(pd.Series(text))

#Save pipeline for later

with open('hbr_biz_pipeline.pkl','wb') as f:
    pickle.dump(test_pipe,f)
