==> 5_Apply model to text.ipynb <==
import pandas as pd
from sklearn import linear_model
from sklearn.externals import joblib
from sklearn.feature_extraction.text import CountVectorizer

# Define a function to take text and score it using the model

def predictor(text):
    #Takes a string, returns prediction of whether it's HBR-relevant
    pipeline = joblib.load('hbr_pipeline.pkl')
    pipeline2 = joblib.load('hbr_biz_pipeline.pkl')
    return pipeline.predict(pd.Series([text]))[0], pipeline2.predict(pd.Series([text]))[0]
    

#Try any text and see what the model says
tk = 'Tesla is having major supply chain problems'
predictor(tk)

#Categorize results form an RSS feed as HBR-relevant or not

import feedparser
feed = 'http://www.nber.org/rss/new.xml'
#feed = 'http://feeds.hbr.org/harvardbusiness'
#feed = 'https://news.ycombinator.com/rss'
#feed = 'https://theconversation.com/us/articles.atom'
feed = feedparser.parse(feed)
for a in feed.entries:
    title = a.title
    print str(predictor(title)) + ": " + title
